+++ packages/taro-fast-component/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 3.0.46
+
+### Patch Changes
+
+- update
+
 ## 3.0.45
 
 ### Patch Changes