+++ packages/taro-fast-framework/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.1.40
+
+### Patch Changes
+
+- Updated dependencies
+  - taro-fast-component@3.0.46
+
 ## 3.1.39
 
 ### Patch Changes